Add docker_images template to sriov overcloud_deploy script
While overcloud deploy is working properly in default deployment
scenario there is an issue when we are trying to deploy with
updating container images. If this template is not present in
the overcloud_deploy script then always default version of
container images are used even if patched container images were
generated. Therefore it is necessary to have this template
specified.

Change-Id: I16eff8a5a8998b0aa1402589b85869fac9e7d327