Add support for trunk_ext workload for migration
This patch adds the following to OVN migration plugin:

* support for trunk_ext workload for OVN migration
infrared plugin. When this workload specified, VMs with trunk ports
will be launched connected to the external network.

* a new fix_workload_mtu flag which is disabled by defualt.
When this flag is enabled each VM MTU will be decreased before OVN
migration. This can be used as a workaround until [1] is fixed.

* added proper trunk ports cleanup into delete_resources template,
plus some cosmetic fixes to delete resources logging.

* added timeouts to ssh commands in order to not get stuck in
password prompt in case server_ssh_user not correct.

* removed old commented code from create_resources template.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1955029

Change-Id: Ieee0bef84f0954ff4f0cf49fa6978eed9a8761c0