Add 'geneve' flag to OVN migration plugin
This patch adds a new flag 'geneve' to OVN migration plugin.
The flag is set to True by default which fits most of the
downstream ovs2ovn CI jobs.
In case user wants target environment to keep using existing
tunnel types, e.g. vlan or vxlan, the flag should be set to
False.
Additionally added a validation that network type is geneve
after the migration in order to not miss migration bugs.

Change-Id: Iaa23db56da290fa15d00106d40b58f077433317b