Fix OVN migration workload creation order
Current order of workload creation is problematic. When we launch VMs
before attaching subnet to router in some cases this causes metadata
issues on one of the VMs. This patch changes the order that should fix
the issue. Now VMs are launched after router is created and subnet is
attached to it.

Change-Id: I7c56b85202e8742adead6d846eb87120d0bf5d97