Add support for Octavia workload to OVN migration plugin
This patch introduces --create_loadbalancer parameter to OVN migration
infrared plugin. When this parameter is set to True (default is False)
OVN migration infrared plugin will create a load balancer and workload
VMs will be added to it as members.
After OVN migration is completed a validation script will check if load
balancer status is correct.
Also a cleanup code for related objects was added.

Change-Id: If3a5c9899575efa7ae06cac79d0753d6fc6970f6