Add option to specify flavor for OVN migration
In some cases we need to specify flavor for OVN migration workload,
e.g. for NFV cases.
Added also an option to skip exporting image, server_user_name and
flavor to OVN migration sctipt in order to allow it to use it's own
values. This can be useful e.g. for SR-IOV job where we use RHEL image
while OVN migration script currently uses a hard-coded small flavor, see [1].
So until [1] is ready we can use the export_image option as a workaround.

[1] https://review.opendev.org/c/openstack/neutron/+/813972

Change-Id: I7837805fe3fef4131afaa82a55f89fb7ca7cf35a