Add option to print VM console output on failure
OVN migration may fail due to connectivity issues between
undercloud and workload VMs. The patch [1] added code for
printing VM console logs for the workload VMs created by
the infrared plugin. This patch adds an option to inject
the code for printing console logs of the VMs created by
the upstream OVN migration code. The option is enabled
by default. This is a temporary workaround
until the VM console printing code will be available in
the upstream rpm.

[1] https://code.engineering.redhat.com/gerrit/c/Neutron-QE/+/310965

Change-Id: I13a41b2b6539bc0c203db93c5b33e98dbc9cdc46