Switch firewall and migrate VMs before ovs2ovn
This patch adds an option to switch firewall driver before ovs2ovn
migration with additional VM migration from a hybrid connection
(through a linux bridge) to a direct connection.
This allows to cover an important scenario that most of ML2/OVS
customers will do if they are going to migrate to OVN. This is
because iptabled_hybrid is the default firewall driver for ML2/OVS
while ovs2ovn migration is not supported in this mode.

Change-Id: I1c45fd9ecbd0d8980447020008bdc0b2a3c5476f