Set valid mtu for trunks
In case ovn migration plugin tries to create trunk on a network with
smaller mtu while by default networks are created with bigger mtu
we get get into a situation when trunk creation can fail due to
invalid mtu. This patch addresses this issue. Now we check mtu of
the network where trunk is connected and create new network with
this mtu.

Change-Id: Ib8371dd95b73c91fa05efb5f83086662a089282e