Add option to disable reducing MTU
At this moment we reduce MTU before starting OVS to OVN migration.
Actually in case tenant networks are VLAN rather than VxLAN it is
not needed. Also we need this option in order to be able to test
some specific customer cases when MTU is already set to a rather
big value and is required to keep not changed when moving from
VxLAN to GENEVE.

Change-Id: I0361b2821f63665ed638ba0de8cd99a449388300