Set lower value for overlay jumbo mtu
In order to cover scenario with keeping overlay MTU unchanged
during ovs2ovn migration we need to set path_mtu value to a lower
value. The value of 8550 was taken from Verizon requirements
document [1]. Also the doc states that "no further MTU size change
is expected to be required as a result of the migration to OVN
because of the switch to the use of Geneve."
With path_mtu set to 8550 all vxlan networks created
with ml2ovs will have MTU of 8500. When migrating to OVN we can skip
executing the reduce-mtu playbook by specifying '--reduce_mtu False'
to the infrared OVN migration plugin and keep using MTU of 8500 also
with Geneve. Ans change of MTU for existing VMs will be not needed.

[1] https://docs.google.com/document/d/1H5C7hI-NlUOqPuYwnTY-_r7R9MS3liGN/edit#heading=h.gjdgxs

Change-Id: I918205622c419e2cd9af3e80aff2538b8421f4a9