Add trunk_ext_with_normal_int workload
There is a need to have a workload that supports ipv6 connectivity
test. Since ipv6 on external network still requires some additional
investigation in order to make it stable, we'll use for now VMs
connected to internal ipv6 subnet.
This patch also sets this workload type as a default one.

Change-Id: Iad268c7db66a920e03ccf6bd6c30aa73b4e0abd1