Run network validation when trunk_ext_with_normal_int
When using trunk_ext_with_normal_int workload type the network
validation was skipped since the workload has 'ext' in the name.
This patch ensures that network validation will run when using
this type of workload since there is an internal network that
is expected to migrate from 'vxlan' to 'geneve'.

Change-Id: Ie56931f05dfc6598eed39ce29f767580cb53979b