Workaround SELinux issue in OVN FDP jobs
(1)
A failure which relates to SELinux permissions was encountered
in past SR-IOV job runs, when latest FDP repo is used,
therefore workarounds were made [0][1].

It was noticed again in composable vlan provider network OVN job,
therefore adding the workaround to other relevant OVN job templates.

(2)
In all job types, the workaround will be triggered when additional repo
is configured for the job run (for example FDP repo).

(3)
A condition for another workaround was removed, this condition is no longer relevant.

[0] https://code.engineering.redhat.com/gerrit/c/nfv-qe/+/428742
[1] https://code.engineering.redhat.com/gerrit/c/Neutron-QE/+/429079

Change-Id: I08c142ab64a83a34951580ae960dd82514625baa