Allow setting a custom value for /srv LV size
With default value of 10 we have issues with restoring overcloud
controller nodes using ReaR tool, see [1] for details.
This patch allows /srv volume size to be configurable
based on change from [2]. This will add an ability to workaround
the issues with restoring controller nodes in order to unblock
testing of OVN migration revert workflow.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2155253
[2] https://review.gerrithub.io/c/redhat-openstack/infrared/+/537498

Change-Id: Ic6dad84d9da4cdbb68ed71a4d510e442faee805c