Don't use controllers as networkers on composable
As appeared, OVN composable roles jobs are scheduling neutron routers
on controller nodes. While networker nodes are also used as gateway
chassis (with lower priority) this is not what these environments
are expected to do. We need to have all routers only on standalone
networker nodes on such environments in order to distinguish between
a usual HA environment, without standalone networkers. Also, some of
l3ha tests rely on the fact that routers are scheduled only on
standelone networker nodes and not on controller nodes.

Change-Id: Ic87c2feb26b1c978743881cade74cb223488bbe7