Add workload suitable for granular migration testing
The initial commit of this workload. It's recommended to have at least
4 compute nodes. In this case selecting --resources_type granular_poc
will create a VM on each compute node. Each VM is connected to external
network. All VMs additionaly have ports connected to a private network,
but there are 2 private networks. VMs are divided to 2 groups and each
group uses own private network. With this setup we can migrate only
compute nodes holding VMs of one group without affecting traffic between
the VMs of this group or other group.

Change-Id: Icb5b1624747b41dc10e91697e2d1c7c378f0beff