Extend workload for granular migration testing
This patch modifies granular_poc workload to provide 2 availability
zones with 2 compute nodes in each zone. 2 VMs on each compute node,
2 networks in each availability zone, each VM is connected to external
network and at least one internal network. There are VMs with trunk
ports in each availability zone, and these VMs are connected to external
and both internal networks of this zone. In general, the workload is
according to this doc [1]. Only networks and VMs names are adjusted
to be aligned with other available migration workloads but the overall
structure and connections are according to [1].

[1] https://docs.google.com/presentation/d/1R6owehStKTLZNeajgubZ7l-WtXt19xa37WFZo1VUkTk/edit#slide=id.g2090297817c_0_0

Change-Id: I8cbb49f71380b22029ca2fa6bdefaa1f6a477d36