Add option to override patchset of batch migration
This is a temporary workaround that will be used for cases when need
to override the default/latest patchset of [1] in case it's not good.
The default value set to 23/440623/4 since it was stable while
with patchset 8 there were failures during deleting neutron resources.
After the batched migration patch [1] is merged we need to remove the
workaround and the code that cherry-picks the neutron patch.

[1] https://code.engineering.redhat.com/gerrit/c/neutron/+/440623

Change-Id: I58fdf36a0d0fc72ac1c9f462660e0426ae44fb83