Workload validation tweaks
This patch adds some improvements to workload validations:
- added an option to run simpler pinger validation. For the cases
of revert where rather long downtime is expected and it's enough to
check that workload VMs are responding at the end. The tests with
a strict downtime check and simpler connectivity test will have
different names.
- validate_ipv6 could not be overridden on restore if it was enabled
on ovn migration. Now it's passed as a valiable to validate-pinger
script, not as a template variable.
- added different title for trunk and non-trunk VM migration tests
in connectivity_summary_file in order to distinguish between the
tests.
- added printing of VMs status at the beginning and the end of
workload validations.

Change-Id: Ic9f8145f1f3a43358aea100d7241caa7eadb13fd